feat: add multiplayer match pages, replay list, and osr parser improvements
New routes:
- MultiplayerMatch, MatchList, MatchView: multiplayer lobby/match history
- MpRedirect: /mp/<id> redirect
- ReplayList: public replay browser
- ApReplayList: ap endpoint for replay listing
- GetFirstPlaces: moved from routes/api to api package

New templates:
- matches.html, match.html: multiplayer UI
- replays.html: replay browser
- analyzer/list.html: analyzer score list
- ap/replays.html, ap/ap_replays.html: ap replay pages

Other:
- OsrParser, OsuFileParser: standalone models (used by Analyzer)
- pom.xml, App.java, ShiinaDocs, ShiinaRoute: route registration updates
- Updated BeatmapQuery, ScoreQuery, FullBeatmap, PubSubModels, SyncedAction
- Updated templates: beatmap, score, user, navbar, base, settings, ap
- New/updated assets: biglogo.gif, wynocha.png, favicon.ico, server logo

+++ src/main/java/dev/osunolimits/models/FullBeatmap.java
@@ -32,7 +32,7 @@ public class FullBeatmap {
     @Data
     public class BeatmapScore {
         private int id;
-        private int pp;
+        private float pp;
         private long score;
         private String[] mods;
         private String grade;
